+++ muslim-data/src/androidTest/java/dev/kosrat/muslimdata/AzkarItemTests.kt
@@ -1,17 +1,14 @@
 package dev.kosrat.muslimdata
 
 import android.content.Context
-import androidx.room.Room
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import dev.kosrat.muslimdata.database.MuslimDataDao
-import dev.kosrat.muslimdata.database.MuslimDataDatabase
 import dev.kosrat.muslimdata.models.Language
-import dev.kosrat.muslimdata.models.Language.*
 import dev.kosrat.muslimdata.repository.MuslimRepository
 import kotlinx.coroutines.runBlocking
 import org.junit.After
-import org.junit.Assert.*
+import org.junit.Assert.assertEquals
+import org.junit.Assert.assertNotNull
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -32,27 +29,32 @@ class AzkarItemTests {
 
     @Test
     fun azkarItems_enItems_isCorrect() {
-        testChapterItems(EN, 1, 4)
+        testChapterItems(Language.EN, 1, 4)
     }
 
     @Test
     fun azkarItems_arItems_isCorrect() {
-        testChapterItems(AR, 25, 8)
+        testChapterItems(Language.AR, 25, 8)
     }
 
     @Test
     fun azkarItems_ckbItems_isCorrect() {
-        testChapterItems(CKB, 50, 2)
+        testChapterItems(Language.CKB, 50, 2)
+    }
+
+    @Test
+    fun azkarItems_ckbBadiniItems_isCorrect() {
+        testChapterItems(Language.CKB_BADINI, 50, 2)
     }
 
     @Test
     fun azkarItems_faItems_isCorrect() {
-        testChapterItems(FA, 75, 1)
+        testChapterItems(Language.FA, 75, 1)
     }
 
     @Test
     fun azkarItems_ruItems_isCorrect() {
-        testChapterItems(RU, 100, 1)
+        testChapterItems(Language.RU, 100, 1)
     }
 
     private fun testChapterItems(language: Language, id: Int, total: Int) = runBlocking {