+++ muslim-data/src/main/assets/database/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the MuslimData database will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.2.0] - 2024-10-30 
+Add Kurdish Badini dialect to the library. 
+
+### Added 
+
+- Add a new Kurdish dialect (Badini dialect as `ckb_BADINI` locale) to the MuslimData library.
+
 ## [2.1.0] - 2024-06-20
 Change comma style in the Arabic azkars to be compatible with the Uthmanic HAFS font.